requirements.txt
SQLAlchemy~=2.0.23
python-dotenv~=1.0.0
mysqlclient~=2.2.0
pydantic~=2.5.1
fastapi~=0.104.1
uvicorn~=0.24.0